Update SDK directory and improve drawing settings panel
This commit updates SDK directory path to correct the build environment. In addition, it includes a major revamp of the drawings settings panel in the application. The changes allow more control over the appearance of drawings by fine-tuning different parameters like the radius of the orbits, steps for the inner and outer orbits, the randomness factor, and the radii of the inner and outer orbit circles. It uses sliders and switches for greater user flexibility. Space has been made for future features like enumerated constraints. Finally, Kotlin and Compose versions have been upgraded for improved functionality and performance.